Early version with support for templating out the NSIS script
This version essentially allows us to avoid manually adding
aircraft/pilots to the NSIS script.

After we download the source files from google drives, we
now parse the files to collect the information about
aircraft and pilots.

With this information, we then jinja template the NSIS script
which then runs to build the installer

* Google Drive secrets are now expected on disk as gdrive_secret.yml
* Increased installer build verbosity in github actions
* Pipenv: Added PyYaml and Jinja
* Added doc/devnotes.md with initial help info
* rs-skin-downloader.py: Cleaned up the google-drive download code
* rs-skin-downloader.py: Improved concurrency with Ray
* rs-skin-downloader.py: Parser of airframes/pilots in Staging
* rs-skin-downloader.py: Templating of the NSI script at the end
* rs-skins-pilot-priorities.ps1:   Not yet implemented
* rs-skins-streamer-priorites.ps1: Not yet implemented
* Removed rs-skins.nsi as this is now rs-skins.jsi.j2 template