Pilot update script is now explicit
We moved from the "intelligent" to "dumb but safe" mode of updating
pilot priorities.

The script used to parse the liveries directory structure and anchor
based on terms like "RED STAR" and $PILOTNAME. This is not
sufficiently "smart", so now we explicitly render out only
the red-start directories to be checked against pilot names.

* Template rs-liveries-pilot-priorities.ps1 instead of
    being "intelligent"
* Added the psexec.nsh NSIS macroes so we can see the powershell
    output in the installer log
* A few variables renamed and removed for clarity